To formate the listing variables and labels For example, if "AESER" and "AEREF" are selected as the listing variable Then, "Intensity" and "Related" can serve as listing label

To formate the listing variables and labels For example, if "AESER" and "AEREF" are selected as the listing variable Then, "Intensity" and "Related" can serve as listing label

define_ae_listing(listing_var = NULL, listing_label = NULL)

Arguments

listing_var
a character vector specifying the variable names to be displayed in the listed details For example, 1) listing_var = c("USUBJID", "SEX") 2) listing_var = c("USUBJID", "SEX", "ADURN", "ADURU")

listing_label
a string vector to specify the according AE labels to be displayed in the selected list. Please note that, if variables contribute to the same listing, their label should be the same. For example, when listing_var = c("ADURN", "ADURU"), listing_label should be listing_label = c("Duration", "Duration"),

Value

a list including (1) the listing_var (3) the listing_label

Examples

define_ae_listing(listing_var = c("USUBJID", "SITEID", "SEX", "RACE", "AGE"),
                  listing_label = c("ID", "Site Number", "Sex", "Race", "Age"))
#> $listing_var
#> [1] "USUBJID" "SITEID"  "SEX"     "RACE"    "AGE"    
#> 
#> $listing_label
#> [1] "ID"          "Site Number" "Sex"         "Race"        "Age"        
#> 
define_ae_listing(listing_var = c("USUBJID", "SEX", "ADURN", "ADURU"),
                  listing_label = c("ID", "Gender", "Duration", "Duration"))
#> $listing_var
#> [1] "USUBJID" "SEX"     "ADURN"   "ADURU"  
#> 
#> $listing_label
#> [1] "ID"       "Gender"   "Duration" "Duration"
#> 
define_ae_listing(listing_var = c("USUBJID", "ADURN", "ADURU", "AGE"),
                  listing_label = c("ID", "Duration", "Duration", "Age"))                           
#> $listing_var
#> [1] "USUBJID" "ADURN"   "ADURU"   "AGE"    
#> 
#> $listing_label
#> [1] "ID"       "Duration" "Duration" "Age"     
#>