sparkline_errorbar

This function is used to draw error bars based on the parameters you defined.

sparkline_errorbar(
  est,
  lower,
  upper,
  color = NULL,
  height = 30,
  xlim = NULL,
  vline = NULL,
  margin = list(l = 0, r = 0, b = 0, t = 0, pad = 0),
  ...
)

Arguments

est
A vector determined the point of your bar

lower
A vector determined the leftmost point

upper
A vector determined the rightmost point

color
A vector of colors you want your bar to be

height
A variable determines the distance between your bars

xlim
Numeric vectors of length 2, giving the x coordinates ranges.

vline
Numeric value for the location to draw the vertical reference line.

margin
Numeric vector of figure margin

...
Additional arguments transfer to sparkline_layout

Value

Sparkline error bar for interactive graphics

Examples

library(plotly)
library(ggplot2)
library(plotly)
est <- c(0,1,2)
lower <- c(-1, -2, -1)
upper <- c(3, 5, 5)
color <- c("red", "blue", "gold")
sparkline_errorbar(est, lower, upper, color)