sparkline_legend

This function is a internal function adding the title to the plot.

sparkline_legend(p, title = "Treatment", pos = -1)

Arguments

p
A parameter indicate which plot to add a tittle

title
A string adding as the tittle of the plot

pos
Numeric value of legend position

Examples

library(plotly)
p <- plot_ly(cars, x = ~speed, y = ~dist)
sparkline_legend(p, title = "cars plot")

#> No trace type specified:
#>   Based on info supplied, a 'scatter' trace seems appropriate.
#>   Read more about this trace type -> https://plotly.com/r/reference/#scatter

#> No scatter mode specifed:
#>   Setting the mode to markers
#>   Read more about this attribute -> https://plotly.com/r/reference/#scatter-mode