Sparkline Point Figure Javascript Render Using Plotly

Sparkline Point Figure Javascript Render Using Plotly

sparkline_point_js(
  tbl,
  x,
  x_lower = NULL,
  x_upper = x_lower,
  xlim = NULL,
  y = seq_along(x),
  vline = NULL,
  text = NULL,
  height = 30,
  width = 150,
  color = "gold",
  color_errorbar = color,
  color_vline = "#00000050"
)

Arguments

tbl
A data frame

x
A vector of variable names in tbl for value.

x_lower
A vector of variable names in tbl for lower error difference.

x_upper
A vector of variable names in tbl for upper error difference. Default is the same as x_lower.

xlim
Numeric vectors of length 2, giving the x coordinates ranges.

y
Numeric vector of y-axis value.

vline
Numeric value for the location to draw the vertical reference line.

text
Character vector of text information dispalted for each point.

height
Numeric value of figure height.

width
Numeric value of figure width.

color
Character vector of point color name

color_errorbar
Character vector of errorbar color name

color_vline
Character vector of vertical reference line color name

Examples

library(reactable)
library(htmltools)

js <- sparkline_point_js(iris, "Sepal.Length")
js <- sparkline_point_js(iris, "Sepal.Length", text = "x")
js <- sparkline_point_js(iris, "Sepal.Length", text = "'count:' + x")
js <- sparkline_point_js(iris, "Sepal.Length", x_lower = "Sepal.Width", text = "'range:' + x + '(' + x_lower + ',' + x_upper + ')'" )
js <- sparkline_point_js(iris, "Sepal.Length", x_lower = "Sepal.Width")
js <- sparkline_point_js(iris, "Sepal.Length", vline = 6)
cat(js)

#> function(cell, state) {
#>  const Plot = createPlotlyComponent(Plotly);
#>  var x = [cell.row['Sepal.Length']];
#>  var y = [1];
#>  var x_lower = [0];
#>  var x_upper = [0];
#>  var x_range = [3.8,8.4];
#>  var y_range = [0,2];
#>  var vline = 6;
#>  var text = "";
#>  var height = 30;
#>  var width = 150;
#>  var color = ['rgba(255,215,0,1)'];
#>  var color_errorbar = 'rgba(255,255,255,0)';
#>  var color_vline = 'rgba(0,0,0,0.313725490196078)';
#> 
#>  return React.createElement(Plot, {
#>     data: [
#>       {
#>        "x": x,
#>        "y": y,
#>        "error_x": {
#>           type: "data",
#>           symmetric: false,
#>           array: x_upper,
#>           arrayminus: x_lower,
#>           "color": color_errorbar
#>        },
#>        "text": text,
#>        "hoverinfo": "text",
#>        "mode": "markers",
#>        "alpha_stroke": 1,
#>        "sizes": [10, 100],
#>        "spans": [1, 20],
#>        "type": "scatter",
#>        "marker": {
#>             "color": color
#>         }
#>       }
#>     ],
#>   "layout": {
#>     "height": height,
#>     "width": width,
#>     "margin": {
#>       "b": 0,
#>       "l": 0,
#>       "t": 0,
#>       "r": 0,
#>       "pad": 0
#>     },
#>     "xaxis": {
#>       "domain": [0,1],
#>       "title": "",
#>       "range": x_range,
#>       "zeroline": false,
#>       "fixedrange": true
#>     },
#>     "yaxis": {
#>       "domain": [0,1],
#>       "title": "",
#>       "range": y_range,
#>       "showgrid": false,
#>       "zeroline": false,
#>       "showticklabels": false,
#>       "fixedrange": true
#>     },
#>     "shapes": [
#>       {
#>         "type": "line",
#>         "y0": y_range[0],
#>         "y1": y_range[1],
#>         "yref": "paper",
#>         "x0": vline,
#>         "x1": vline,
#>         "line": {
#>           "color": color_vline
#>         }
#>       }
#>     ],
#>     "plot_bgcolor": "rgba(0, 0, 0, 0)",
#>     "paper_bgcolor": "rgba(0, 0, 0, 0)",
#>     "hoverlabel": {
#>       "bgcolor": "lightgray"
#>     },
#>     "showlegend": false,
#>     "hovermode": "closest"
#>   },
#>   "config": {
#>     "showSendToCloud": false,
#>     "displayModeBar": false
#>   }
#> 
#> 
#>   })
#> }

p <- reactable(data = iris[,1:2],
               theme = reactableTheme(cellPadding = "0px 8px"),
               borderless = TRUE,
               highlight = TRUE,
               resizable = TRUE,
               columns = list( Sepal.Length = colDef(html = TRUE,
                                                     cell = JS(js),
                                                     width = 150,
                                                     style="font-size: 0px; padding: 0px; margin: 0px;")))

browsable(tagList(
  reactR::html_dependency_react(),
  htmltools::tags$script(src="https://cdn.plot.ly/plotly-latest.min.js"),
  htmltools::tags$script(src="https://unpkg.com/react-plotly.js@1.0.2/dist/create-plotly-component.js"),
  p
))