tidy_ae_table

Define analysis data This function is used to obtain AE information ready for visualization

tidy_ae_table(
  population_from,
  observation_from,
  population_where = "ITTFL=='Y'",
  observation_where = NULL,
  treatment_var = "TRTA",
  treatment_order = unique(population_from[[treatment_var]]),
  ae_var = "AEDECOD",
  stratum_var = NULL,
  listing_var = names(observation_from)
)

Arguments

population_from
A data frame to obtain population level variables.Typically an 'adsl' dataset.

observation_from
A data frame to obtain observation level variables.

population_where
A character string to define the criteria to select analysis population.

observation_where
A character string to define the criteria to select analysis observation.

treatment_var
A character string to define the variable of new column called "treatment"

treatment_order
A character vector to define the treatment display order and label.

ae_var
A character string to define the variable of new column called ae"

stratum_var
A character string to define the variable of baseline stratum in 'population_from'.Only one 'stratum_var' is allowed.

listing_var
A character string to define the criteria to select the column of the table

Value

Return a standard adverse event data frame

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(tidyr)
db <- tidy_ae_table(population_from  = adsl %>% rename(TRTA = TRT01A),
                    observation_from = adae,
                    treatment_var = "TRTA",
                    treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                    listing_var = c("SITEID", "USUBJID", "AGE", "RACE", "SEX",
                                    "AETERM", "AESER", "AEREL", "AEACN", "AEOUT") )

#> Joining, by = "treatment"