This is a function to rbind data together according to the interested ae labels

This is a function to rbind data together according to the interested ae labels

tidy_multi_ae_label(db, db_N, ae_interested = NULL)

Arguments

db
a data frame which is usually an intersection of adae and adsl, with the ae variable name (usually encoded as "AEDECOD") transformed into "ae", and treatment variable name (usually encoded as "TRTA") transformed into "treatment", see the following example for the format of db

db_N
a data frame contains the sample size of each arm

ae_interested
A character string including the interested AE structure, ex., AESER, AEREL...

Value

a data frame which rbind data together according to the interested ae labels

Examples

library(dplyr)
treatment_order = c("MK9999" = "Xanomeline", "Placebo" = "Placebo")

db = tibble(USUBJID = c("01", "01", "02", "03", "03", "03", "04", "04", "05"),
            stratum = c("a", "a", "b", "c", "c", "c", "b", "b", "a"),
            treatment = factor(c("Xanomeline", "Xanomeline", "Placebo", "Placebo", "Placebo", 
                                 "Placebo", "Xanomeline", "Xanomeline", "Xanomeline"), 
                               levels = treatment_order, labels = treatment_order),
            ae = c("headache", "pain", "headache", "fever", "running nose", 
                           "pain", "headache", "fever", "headache"),
            AESER = c("N", "N", "N", "Y", "N", "N", "N", "Y", "N"),
            AEREL = c("N", "Y", "N", "N", "N", "Y", "N", "N", "Y"))
                    
db_N = tibble(treatment = factor(c("Xanomeline", "Placebo"), 
                                 levels = treatment_order, 
                                 labels = treatment_order), 
              N = c(3, 2))

tidy_multi_ae_label(db, db_N, ae_interested = NULL)

#> Joining, by = "treatment"

#> # A tibble: 12 × 9
#>    ae           stratum ae_label   n_1   n_2   N_1   N_2 pct_1 pct_2
#>    <chr>        <chr>   <chr>    <int> <int> <dbl> <dbl> <dbl> <dbl>
#>  1 fever        a       All          0     0     3     2   0       0
#>  2 fever        b       All          1     0     3     2  33.3     0
#>  3 fever        c       All          0     1     3     2   0      50
#>  4 headache     a       All          2     0     3     2  66.7     0
#>  5 headache     b       All          1     1     3     2  33.3    50
#>  6 headache     c       All          0     0     3     2   0       0
#>  7 pain         a       All          1     0     3     2  33.3     0
#>  8 pain         b       All          0     0     3     2   0       0
#>  9 pain         c       All          0     1     3     2   0      50
#> 10 running nose a       All          0     0     3     2   0       0
#> 11 running nose b       All          0     0     3     2   0       0
#> 12 running nose c       All          0     1     3     2   0      50

tidy_multi_ae_label(db, db_N, ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"', 'AEREL != "N"'),
                                                                    ae_label = c("with serious adverse events","with drug-related adverse events")))

#> Joining, by = "treatment"

#> Joining, by = "treatment"

#> Joining, by = "treatment"

#> # A tibble: 15 × 9
#>    ae           stratum ae_label               n_1   n_2   N_1   N_2 pct_1 pct_2
#>    <chr>        <chr>   <chr>                <int> <int> <dbl> <dbl> <dbl> <dbl>
#>  1 fever        a       All                      0     0     3     2   0       0
#>  2 fever        b       All                      1     0     3     2  33.3     0
#>  3 fever        c       All                      0     1     3     2   0      50
#>  4 headache     a       All                      2     0     3     2  66.7     0
#>  5 headache     b       All                      1     1     3     2  33.3    50
#>  6 headache     c       All                      0     0     3     2   0       0
#>  7 pain         a       All                      1     0     3     2  33.3     0
#>  8 pain         b       All                      0     0     3     2   0       0
#>  9 pain         c       All                      0     1     3     2   0      50
#> 10 running nose a       All                      0     0     3     2   0       0
#> 11 running nose b       All                      0     0     3     2   0       0
#> 12 running nose c       All                      0     1     3     2   0      50
#> 13 fever        NA      with serious advers…     1     1     3     2  33.3    50
#> 14 headache     NA      with drug-related a…     1     0     3     2  33.3     0
#> 15 pain         NA      with drug-related a…     1     1     3     2  33.3    50