tidy_population

This function is used to prepare the observation datasets for the use in function tidy_ae_table(). The function filter the adsl dataset, and define proper treatment order in the dataset.

tidy_population(
  population_from,
  population_where = NULL,
  treatment_var = "TRT01A",
  treatment_order = unique(population_from[[treatment_var]]),
  stratum_var = NULL,
  baseline_var = NULL
)

Arguments

population_from
A data frame to obtain population level variables. Typically an 'adsl' dataset. It is the source of variables mentioned in population_where, treatment_var, stratum_var and covariate_var.

population_where
A character string to define the criteria to select analysis population.

treatment_var
A character string to define the variable of new column called "treatment".

treatment_order
A vector of character strings that tells the function which rows of the table should be select, only if the values in "treatment" is in this vector. It also provide the label names of the treatments after turning them into factors.

stratum_var
A character string to define the variable of baseline stratum in 'population_from'. Only one 'stratum_var' is allowed.

baseline_var
Tells the function which column of the table to be selected as the final output.

Value

an output data frame of population level information. USUBJID: Unique subject identifier. treatment: Actual treatment. stratum: baseline stratum in 'population_from'.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

pop <- tidy_population(population_from  = adsl %>% rename(TRTA = TRT01A),
 population_where = "ITTFL=='Y'",
 treatment_var    = "TRTA",
 treatment_order  = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
 stratum_var      = NULL,
 baseline_var     = NULL)

tidy_population(population_from = adsl)

#> # A tibble: 254 × 3
#>    USUBJID     treatment            stratum
#>    <chr>       <fct>                <chr>  
#>  1 01-701-1015 Placebo              All    
#>  2 01-701-1023 Placebo              All    
#>  3 01-701-1028 Xanomeline High Dose All    
#>  4 01-701-1033 Xanomeline Low Dose  All    
#>  5 01-701-1034 Xanomeline High Dose All    
#>  6 01-701-1047 Placebo              All    
#>  7 01-701-1097 Xanomeline Low Dose  All    
#>  8 01-701-1111 Xanomeline Low Dose  All    
#>  9 01-701-1115 Xanomeline Low Dose  All    
#> 10 01-701-1118 Placebo              All    
#> # … with 244 more rows

tidy_population(population_from = adsl, stratum_var = "RACE")

#> # A tibble: 254 × 3
#>    USUBJID     treatment            stratum
#>    <chr>       <fct>                <chr>  
#>  1 01-701-1015 Placebo              WHITE  
#>  2 01-701-1023 Placebo              WHITE  
#>  3 01-701-1028 Xanomeline High Dose WHITE  
#>  4 01-701-1033 Xanomeline Low Dose  WHITE  
#>  5 01-701-1034 Xanomeline High Dose WHITE  
#>  6 01-701-1047 Placebo              WHITE  
#>  7 01-701-1097 Xanomeline Low Dose  WHITE  
#>  8 01-701-1111 Xanomeline Low Dose  WHITE  
#>  9 01-701-1115 Xanomeline Low Dose  WHITE  
#> 10 01-701-1118 Placebo              WHITE  
#> # … with 244 more rows