tlf_ae_summary

This function is used to generate AE summary table

tlf_ae_summary(
  population_from,
  observation_from,
  population_where,
  observation_where,
  treatment_var,
  treatment_order,
  ae_var,
  ae_interested = NULL,
  stratum_var = NULL,
  display_ci = FALSE,
  display_total = FALSE,
  title_text,
  subtitle_text = NULL,
  end_notes,
  output_name = "./tlf_ae_summary.rtf"
)

Arguments

population_from
A data frame to obtain population level variables.Typically an 'adsl' dataset.

observation_from
A data frame to obtain observation level variables.

population_where
A character string to define the criteria to select analysis population.

observation_where
A character string to define the criteria to select analysis observation.

treatment_var
A character string to define the variable of new column called "treatment"

treatment_order
A character vector to define the treatment display order and label.

ae_var
A character string to define the variable of new column called ae

ae_interested
An object returned by function define_ae_select_list()

stratum_var
A character string to define the variable of baseline stratum in 'population_from'.Only one 'stratum_var' is allowed.

display_ci
A logic value indicating whether displaying the risk difference, corresponding confidence interval and its p-value. Default is FALSE.

display_total
A logic value indicating whether displaying the total column. Default is FALSE.

title_text
The part of the text string to appear in the title row.

subtitle_text
A vector of text strings to appear in the subtitle row(s).

end_notes
A vector of text strings to appear in the footnote.

output_name
A text string to output the table, can be specified with the path together. If no path specified, the default path will be used to save the table.

Value

Return a table for summary of AE

Examples

library(dplyr)
library(r2rtf)
tlf_ae_summary(population_from = adsl %>% rename(TRTA = TRT01A),
                observation_from = adae,
                population_where = "ITTFL=='Y'",
                observation_where = NULL,
                treatment_var = "TRTA",
                treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                ae_var = "AEDECOD",
                ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"', 'AEREL != "NONE"',
                                                                       'AEACN == "DRUG WITHDRAWN"'),
                                                      ae_label = c("with serious adverse events",
                                                                   "with drug-related adverse events",
                                                                   "discontinued due to an adverse event")),
                stratum_var = NULL,
                display_ci = TRUE,
                display_total = FALSE,
                title_text = "Analysis of Adverse Event Summary", 
                subtitle_text = NULL,
                end_notes ="Every subject is counted a single time for each applicable row and column.",
                output_name = file.path(tempdir(), 'ae0summary.rtf'))

#> Joining, by = "treatment"

#> Joining, by = "treatment"

#> Joining, by = "treatment"

#> Warning: NaNs produced

#> Warning: NaNs produced