Display interactive forest plot

Display interactive forest plot

Usage

ae_forestly(
  outdata,
  display_soc_toggle = TRUE,
  filter = c("prop", "n"),
  filter_label = NULL,
  filter_range = NULL,
  width = 1400,
  max_page = NULL
)

Arguments

outdata

An outdata object created by format_ae_forestly().

display_soc_toggle

A boolean value to display SOC toggle button.

filter

A character value of the filter variable.

filter_label

A character value of the label for slider bar.

filter_range

A numeric vector of length 2 for the range of the slider bar. If NULL (default), the range is automatically calculated from the data. If only one value is provided, it will be used as the maximum and minimum will be 0.

width

A numeric value of width of the table in pixels.

max_page

A numeric value of max page number shown in the table.

Value

An AE forest plot saved as a shiny.tag.list object.

Examples

adsl <- forestly_adsl[1:100, ]
adae <- forestly_adae[1:100, ]
if (interactive()) {
  meta_forestly(
    dataset_adsl = adsl,
    dataset_adae = adae,
  ) |>
    prepare_ae_forestly() |>
    format_ae_forestly() |>
    ae_forestly()
}