Generate interactive forest plot with a select list as filter

Generate interactive forest plot with a select list as filter

forestly(
  db,
  fig_prop_color = c("#00857C", "#66203A"),
  fig_prop_label = NULL,
  fig_prop_colwidth = 300,
  fig_diff_color = "black",
  fig_diff_label = NULL,
  fig_diff_colwidth = 300,
  small_sample = NULL,
  forest_table_width = 1200,
  forest_table_bottom_margin = 50
)

Arguments

db
a list of two components: (1) db$table: ae data frame rbinded by interested ae labels (2) db$listing: a standard adverse event data frame with select columns

fig_prop_color
a character vector of length two to specify the colors to plot the two proportions. The first entry refers to the color of the AE proportion in the treatment arm. The second entry refers to the color of the AE proportion in the control arm. The default value is c("gold", "purple").

fig_prop_label
a character vector of length two to specify the x-axis label of the two propositions plots. The first entry refers to the x-lab legend of the treatment arm. The second entry refers to the x-lab legend of the control arm. The default value is c("treatment", "control")

fig_prop_colwidth
a numerical value to specify the column width of the two proportion plot the default value is 300

fig_diff_color
a string to specify the color to plot the error bar. The default value is "black".

fig_diff_label
a string to specify the x-axis legend of the error bar. The default value is "treatment <- Favor -> control".

fig_diff_colwidth
a numerical value to specify the column width of the two proportion plot the default value is 300

small_sample
a integral vector of length 2. The first element is for treatment group and the second element is for control group. The default value is c(4, 4).

forest_table_width
the width of the forest plot reactbale

forest_table_bottom_margin
the bottom margin of the forest plot reactable

Value

a reactable with a select list

Examples

if (FALSE) {
library(forestly)
## No stratum ----------------------------------------------
## Tidy data
db <- tidy_ae_table(population_from  = adsl %>% rename(TRTA = TRT01A),
                    observation_from = adae,
                    population_where = NULL,
                    observation_where = NULL,
                    treatment_var = "TRTA",
                    treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                    stratum_var = NULL,
                    ae_var = "AEDECOD",
                    ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"',
                                                                           'AEREL != "NONE"'),
                                                          ae_label = c("with serious adverse events",
                                                                       "with drug-related adverse events")),
                    listing_interested = define_ae_listing(listing_var = c("USUBJID", "SEX", "AGE", "AEDECOD"),
                                                           listing_label = c("ID", "Gender", "Age", "Adverse Event")),
                    ae_listing_label = "Adverse Event")
## Generate interavtive forest plot
forestly(db)
forestly(db,
         fig_prop_color = c("gold", "purple"),
         fig_prop_label = NULL,
         fig_diff_color = "black",
         fig_diff_label = NULL,
         small_sample = c(4, 4))

## With stratum ----------------------------------------------
## Impute stratum 
adsl$STRATUMN <- sample(seq(1,3), size = length(adsl$USUBJID), prob = c(0.3, 0.3, 0.4), replace = TRUE)
adae <- adae %>% left_join(data.frame(USUBJID = adsl$USUBJID, STRATUMN = adsl$STRATUMN))
## Tidy data
db <- tidy_ae_table(population_from  = adsl %>% rename(TRTA = TRT01A),
                    observation_from = adae,
                    population_where = NULL,
                    observation_where = NULL,
                    treatment_var = "TRTA",
                    treatment_order = c("MK9999" = "Xanomeline High Dose", "Placebo" = "Placebo"),
                    stratum_var = "STRATUMN", 
                    ae_var = "AEDECOD",
                    ae_interested = define_ae_select_list(ae_criterion = c('AESER == "Y"',
                                                                           'AEREL != "NONE"'),
                                                          ae_label = c("with serious adverse events",
                                                                       "with drug-related adverse events")),
                    listing_interested = define_ae_listing(listing_var = c("USUBJID", "SEX", "AGE", "AEDECOD"),
                                                           listing_label = c("ID", "Gender", "Age", "Adverse Event")),
                    ae_listing_label = "Adverse Event")
## Generate interavtive forest plot
forestly(db)
forestly(db,
         fig_prop_color = c("gold", "purple"),
         fig_prop_label = NULL,
         fig_diff_color = "black",
         fig_diff_label = NULL,
         small_sample = c(4, 4))
}